Major async profiler tests refactory, split into v1 and v2.
